refactor: share OTLP proto parsing across generators
